agent sessions - more code cleanup
