+++ src/vs/workbench/contrib/chat/browser/agentSessions/agentSessionsActions.ts
@@ -5,7 +5,7 @@
 
 import { localize, localize2 } from '../../../../../nls.js';
 import { IAgentSession } from './agentSessionsModel.js';
-import { Action2, MenuId } from '../../../../../platform/actions/common/actions.js';
+import { Action2, MenuId, MenuRegistry } from '../../../../../platform/actions/common/actions.js';
 import { Codicon } from '../../../../../base/common/codicons.js';
 import { ServicesAccessor } from '../../../../../editor/browser/editorExtensions.js';
 import { AgentSessionsViewerOrientation, IAgentSessionsControl, IMarshalledChatSessionContext, isMarshalledChatSessionContext } from './agentSessions.js';
@@ -27,6 +27,106 @@ import { ACTION_ID_NEW_CHAT, CHAT_CATEGORY } from '../actions/chatActions.js';
 import { IViewsService } from '../../../../services/views/common/viewsService.js';
 import { ChatViewPane } from '../chatViewPane.js';
 import { ICommandService } from '../../../../../platform/commands/common/commands.js';
+
+//#region Chat View
+
+export class ToggleChatViewSessionsAction extends Action2 {
+	constructor() {
+		super({
+			id: 'workbench.action.chat.toggleChatViewSessions',
+			title: localize2('chat.toggleChatViewSessions.label', "Show Sessions"),
+			toggled: ContextKeyExpr.equals(`config.${ChatConfiguration.ChatViewSessionsEnabled}`, true),
+			menu: {
+				id: MenuId.ChatWelcomeContext,
+				group: '0_sessions',
+				order: 1,
+				when: ChatContextKeys.inChatEditor.negate()
+			}
+		});
+	}
+
+	async run(accessor: ServicesAccessor): Promise<void> {
+		const configurationService = accessor.get(IConfigurationService);
+
+		const chatViewSessionsEnabled = configurationService.getValue<boolean>(ChatConfiguration.ChatViewSessionsEnabled);
+		await configurationService.updateValue(ChatConfiguration.ChatViewSessionsEnabled, !chatViewSessionsEnabled);
+	}
+}
+
+const agentSessionsOrientationSubmenu = new MenuId('chatAgentSessionsOrientationSubmenu');
+MenuRegistry.appendMenuItem(MenuId.ChatWelcomeContext, {
+	submenu: agentSessionsOrientationSubmenu,
+	title: localize('chat.sessionsOrientation', "Sessions Orientation"),
+	group: '0_sessions',
+	order: 2,
+	when: ChatContextKeys.inChatEditor.negate()
+});
+
+export class SetAgentSessionsOrientationAutoAction extends Action2 {
+
+	constructor() {
+		super({
+			id: 'workbench.action.chat.setAgentSessionsOrientationAuto',
+			title: localize2('chat.sessionsOrientation.auto', "Auto"),
+			toggled: ContextKeyExpr.equals(`config.${ChatConfiguration.ChatViewSessionsOrientation}`, 'auto'),
+			precondition: ContextKeyExpr.equals(`config.${ChatConfiguration.ChatViewSessionsEnabled}`, true),
+			menu: {
+				id: agentSessionsOrientationSubmenu,
+				group: 'navigation',
+				order: 1
+			}
+		});
+	}
+
+	async run(accessor: ServicesAccessor): Promise<void> {
+		const configurationService = accessor.get(IConfigurationService);
+		await configurationService.updateValue(ChatConfiguration.ChatViewSessionsOrientation, 'auto');
+	}
+}
+
+export class SetAgentSessionsOrientationStackedAction extends Action2 {
+
+	constructor() {
+		super({
+			id: 'workbench.action.chat.setAgentSessionsOrientationStacked',
+			title: localize2('chat.sessionsOrientation.stacked', "Stacked"),
+			toggled: ContextKeyExpr.equals(`config.${ChatConfiguration.ChatViewSessionsOrientation}`, 'stacked'),
+			precondition: ContextKeyExpr.equals(`config.${ChatConfiguration.ChatViewSessionsEnabled}`, true),
+			menu: {
+				id: agentSessionsOrientationSubmenu,
+				group: 'navigation',
+				order: 2
+			}
+		});
+	}
+
+	async run(accessor: ServicesAccessor): Promise<void> {
+		const configurationService = accessor.get(IConfigurationService);
+		await configurationService.updateValue(ChatConfiguration.ChatViewSessionsOrientation, 'stacked');
+	}
+}
+
+export class SetAgentSessionsOrientationSideBySideAction extends Action2 {
+
+	constructor() {
+		super({
+			id: 'workbench.action.chat.setAgentSessionsOrientationSideBySide',
+			title: localize2('chat.sessionsOrientation.sideBySide', "Side by Side"),
+			toggled: ContextKeyExpr.equals(`config.${ChatConfiguration.ChatViewSessionsOrientation}`, 'sideBySide'),
+			precondition: ContextKeyExpr.equals(`config.${ChatConfiguration.ChatViewSessionsEnabled}`, true),
+			menu: {
+				id: agentSessionsOrientationSubmenu,
+				group: 'navigation',
+				order: 3
+			}
+		});
+	}
+
+	async run(accessor: ServicesAccessor): Promise<void> {
+		const configurationService = accessor.get(IConfigurationService);
+		await configurationService.updateValue(ChatConfiguration.ChatViewSessionsOrientation, 'sideBySide');
+	}
+}
 
 export class FocusAgentSessionsAction extends Action2 {
 
@@ -67,6 +167,10 @@ export class FocusAgentSessionsAction extends Action2 {
 	}
 }
 
+//#endregion
+
+//#region Session Title Actions
+
 abstract class BaseAgentSessionAction extends Action2 {
 
 	run(accessor: ServicesAccessor, context: IAgentSession | IMarshalledChatSessionContext): void {
@@ -86,8 +190,6 @@ abstract class BaseAgentSessionAction extends Action2 {
 
 	abstract runWithSession(session: IAgentSession, accessor: ServicesAccessor): void;
 }
-
-//#region Session Title Actions
 
 export class ArchiveAgentSessionAction extends BaseAgentSessionAction {
