+++ src/vs/workbench/contrib/chat/browser/chatImageUtils.ts
