+++ src/vs/workbench/contrib/chat/browser/chatSessions/media/chatSessionPickerActionItem.css
