+++ src/vs/workbench/contrib/chat/browser/contrib/chatScreenshotContext.ts
