+++ src/vs/workbench/contrib/chat/common/chatSessionsService.ts
@@ -78,9 +78,6 @@ export interface IChatSessionItem {
 		deletions: number;
 	} | readonly IChatSessionFileChange[];
 	archived?: boolean;
-	// TODO:@osortega remove once the single-view is default
-	/** @deprecated */
-	history?: boolean;
 }
 
 export interface IChatSessionFileChange {
@@ -146,11 +143,6 @@ export interface IChatSessionItemProvider {
 	readonly chatSessionType: string;
 	readonly onDidChangeChatSessionItems: Event<void>;
 	provideChatSessionItems(token: CancellationToken): Promise<IChatSessionItem[]>;
-	provideNewChatSessionItem?(options: {
-		request: IChatAgentRequest;
-		// eslint-disable-next-line @typescript-eslint/no-explicit-any
-		metadata?: any;
-	}, token: CancellationToken): Promise<IChatSessionItem>;
 }
 
 export interface IChatSessionContentProvider {
